// src/types/signup.ts
export interface IFormInput {
    email: string;
    otp: number;
    name: string;
    oldPassword: string;
    newPassword: string;
    password: string;
    confirmPassword: string;
  }
  
//   export interface FormDataProps {
//     file: File | null;
//     reviewerLocation: string;
//     reviewerName: string;
//     review: string;
//     isPublished: string;
//   }